arm-optimized-routines: build and run the tests
This doesn't enable presubmit because the tests don't pass on 32-bit.

Strictly, we're also not running the ulp tests because they're a
bloodbath on both LP32 and LP64 at the moment.

But at least the 64-bit non-ulp tests are passing, and this patch is
enough to let you run any/all of the other tests.

Test: atest arm-optimized-routines-tests (slow)
Test: ./run-arm-optimized-routines-tests-on-android.sh (fast)
Change-Id: I07b78e1222064af6877f311ef9bbd76e4a9a3822
Signed-off-by: mydongistiny <jaysonedson@gmail.com>